+++ services/editor/src/components/common/EditableText/EditableText.less
@@ -1,4 +1,4 @@
-@import (reference) "../../../../../../styles/core/core.less";
+@import (reference) "../../../styles/core/core.less";
 .editable-text-container {
   @text-color: #515c66;
   border: 1px solid transparent;
