+++ services/editor/src/components/common/EditableTextArea/EditableTextArea.css
