+++ services/editor/src/components/common/EditableTextArea/EditableTextArea.less
@@ -1,4 +1,4 @@
-@import (reference) "../../../../../../styles/core/core.less";
+@import (reference) "../../../styles/core/core.less";
 .textarea-container {
   .textarea-input {
     display: block;
